+++ PXCTestKit/Extension/FBSimulator+PXCTestKit.swift
@@ -47,6 +47,9 @@ extension Sequence where Iterator.Element == FBSimulator {
     func install(applications: [FBApplicationDescriptor]) throws {
         for simulator in self {
             for application in applications {
+                if simulator.installedApplications.filter({ $0.bundleID == application.bundleID }).count == 1 {
+                    try simulator.interact.uninstallApplication(withBundleID: application.bundleID).perform()
+                }
                 try simulator.interact.installApplication(application).perform()
             }
         }